Minor folder and file renaming
