+++ src/main/java/frc/robot/RobotContainer.java
@@ -22,11 +22,11 @@ import edu.wpi.first.wpilibj2.command.button.JoystickButton;
 import frc.robot.DriverProfiles.ProfilingManagement;
 import frc.robot.commands.swerve.StraightenWheels;
 import frc.robot.commands.swerve.TurnToVision;
-import frc.robot.commands.swerve.drive.LetsRoll1Joystick;
-import frc.robot.commands.swerve.drive.LetsRoll2Joysticks;
-import frc.robot.commands.swerve.test.TestModulePID;
-import frc.robot.commands.swerve.test.TestModuleVelocity;
-import frc.robot.commands.swerve.test.TestMovementPID;
+import frc.robot.commands.swerve.driving.LetsRoll1Joystick;
+import frc.robot.commands.swerve.driving.LetsRoll2Joysticks;
+import frc.robot.commands.swerve.testing.TestModulePID;
+import frc.robot.commands.swerve.testing.TestModuleVelocity;
+import frc.robot.commands.swerve.testing.TestMovementPID;
 import frc.robot.subsystems.NavX;
 import frc.robot.subsystems.Superstructure;
 import frc.robot.subsystems.Vision;
