+++ src/main/java/frc/robot/commands/swerve/driving/DriveForTime.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.swerve.drive;
+package frc.robot.commands.swerve.driving;
 
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.CommandBase;
@@ -10,14 +10,14 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.Parameters;
 import frc.robot.RobotContainer;
 
-public class DriveForTIme extends CommandBase {
+public class DriveForTime extends CommandBase {
     /** Creates a new Drive. */
     Timer timer = new Timer();
 
     double time = 0;
     double speed = 0;
 
-    public DriveForTIme(double time) {
+    public DriveForTime(double time) {
         // Use addRequirements() here to declare subsystem dependencies.
         addRequirements(RobotContainer.driveTrain);
         this.time = time;
